+++ miru/context/view.py
@@ -1,0 +1,49 @@
+from __future__ import annotations
+
+import typing as t
+
+from .raw import RawComponentContext
+
+if t.TYPE_CHECKING:
+    from ..interaction import ComponentInteraction
+    from ..view import View
+
+__all__ = ("ViewContext",)
+
+
+class ViewContext(RawComponentContext):
+    """A context object proxying a ComponentInteraction for a view item."""
+
+    __slots__ = ("_interaction", "_responses", "_view")
+
+    def __init__(self, view: View, interaction: ComponentInteraction) -> None:
+        super().__init__(interaction)
+        self._view = view
+
+    @property
+    def view(self) -> View:
+        """The view this context originates from."""
+        return self._view
+
+
+# MIT License
+#
+# Copyright (c) 2022-present HyperGH
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
