+++ miru/interaction.py
@@ -1,16 +1,11 @@
 from __future__ import annotations
 
-import asyncio
-import datetime
 import functools
 import typing as t
 
 import hikari
 
-if t.TYPE_CHECKING:
-    from .context import Context
-
-__all__ = ["ComponentInteraction", "ModalInteraction", "InteractionResponse"]
+__all__ = ["ComponentInteraction", "ModalInteraction"]
 
 
 class ComponentInteraction(hikari.ComponentInteraction):
@@ -114,146 +109,6 @@ class ModalInteraction(hikari.ModalInteraction):
 
 
 InteractionT = t.TypeVar("InteractionT", "ComponentInteraction", "ModalInteraction")
-
-
-class InteractionResponse:
-    """
-    Represents a response to an interaction, allows for standardized handling of responses.
-    This class is not meant to be directly instantiated, and is instead returned by :obj:`miru.context.Context`.
-    """
-
-    def __init__(self, context: Context[InteractionT], message: t.Optional[hikari.Message] = None) -> None:
-        self._context: Context[t.Any] = context  # Before you ask why it is Any, because mypy is dumb
-        self._message: t.Optional[hikari.Message] = message
-        self._delete_after_task: t.Optional[asyncio.Task[None]] = None
-
-    async def _do_delete_after(self, delay: float) -> None:
-        """Delete the response after the specified delay.
-
-        This should not be called manually,
-        and instead should be triggered by the ``delete_after`` method of this class.
-        """
-        await asyncio.sleep(delay)
-        await self.delete()
-
-    def delete_after(self, delay: t.Union[int, float, datetime.timedelta]) -> None:
-        """Delete the response after the specified delay.
-
-        Parameters
-        ----------
-        delay : Union[int, float, datetime.timedelta]
-            The delay after which the response should be deleted.
-        """
-        if self._delete_after_task is not None:
-            raise RuntimeError("A delete_after task is already running.")
-
-        if isinstance(delay, datetime.timedelta):
-            delay = delay.total_seconds()
-        self._delete_after_task = asyncio.create_task(self._do_delete_after(delay))
-
-    async def retrieve_message(self) -> hikari.Message:
-        """Get or fetch the message created by this response.
-        Initial responses need to be fetched, while followups will be provided directly.
-
-        Returns
-        -------
-        hikari.Message
-            The message created by this response.
-        """
-        if self._message:
-            return self._message
-
-        assert isinstance(self._context.interaction, (ComponentInteraction, ModalInteraction))
-        return await self._context.interaction.fetch_initial_response()
-
-    async def delete(self) -> None:
-        """Delete the response issued to the interaction this object represents."""
-
-        if self._message:
-            await self._context.interaction.delete_message(self._message)
-
-        await self._context.interaction.delete_initial_response()
-
-    async def edit(
-        self,
-        content: hikari.UndefinedOr[t.Any] = hikari.UNDEFINED,
-        *,
-        component: hikari.UndefinedOr[hikari.api.ComponentBuilder] = hikari.UNDEFINED,
-        components: hikari.UndefinedOr[t.Sequence[hikari.api.ComponentBuilder]] = hikari.UNDEFINED,
-        attachment: hikari.UndefinedOr[hikari.Resourceish] = hikari.UNDEFINED,
-        attachments: hikari.UndefinedOr[t.Sequence[hikari.Resourceish]] = hikari.UNDEFINED,
-        embed: hikari.UndefinedOr[hikari.Embed] = hikari.UNDEFINED,
-        embeds: hikari.UndefinedOr[t.Sequence[hikari.Embed]] = hikari.UNDEFINED,
-        replace_attachments: bool = False,
-        mentions_everyone: hikari.UndefinedOr[bool] = hikari.UNDEFINED,
-        user_mentions: hikari.UndefinedOr[
-            t.Union[hikari.SnowflakeishSequence[hikari.PartialUser], bool]
-        ] = hikari.UNDEFINED,
-        role_mentions: hikari.UndefinedOr[
-            t.Union[hikari.SnowflakeishSequence[hikari.PartialRole], bool]
-        ] = hikari.UNDEFINED,
-    ) -> InteractionResponse:
-        """A short-hand method to edit the message belonging to this response.
-
-        Parameters
-        ----------
-        content : undefined.UndefinedOr[t.Any], optional
-            The content of the message. Anything passed here will be cast to str.
-        attachment : undefined.UndefinedOr[hikari.Resourceish], optional
-            An attachment to add to this message.
-        attachments : undefined.UndefinedOr[t.Sequence[hikari.Resourceish]], optional
-            A sequence of attachments to add to this message.
-        component : undefined.UndefinedOr[hikari.api.special_endpoints.ComponentBuilder], optional
-            A component to add to this message.
-        components : undefined.UndefinedOr[t.Sequence[hikari.api.special_endpoints.ComponentBuilder]], optional
-            A sequence of components to add to this message.
-        embed : undefined.UndefinedOr[hikari.Embed], optional
-            An embed to add to this message.
-        embeds : undefined.UndefinedOr[t.Sequence[hikari.Embed]], optional
-            A sequence of embeds to add to this message.
-        mentions_everyone : undefined.UndefinedOr[bool], optional
-            If True, mentioning @everyone will be allowed.
-        user_mentions : undefined.UndefinedOr[t.Union[hikari.SnowflakeishSequence[hikari.PartialUser], bool]], optional
-            The set of allowed user mentions in this message. Set to True to allow all.
-        role_mentions : undefined.UndefinedOr[t.Union[hikari.SnowflakeishSequence[hikari.PartialRole], bool]], optional
-            The set of allowed role mentions in this message. Set to True to allow all.
-
-        Returns
-        -------
-        InteractionResponse
-            A proxy object representing the response to the interaction.
-        """
-        if self._message:
-            message = await self._context.interaction.edit_message(
-                self._message,
-                content,
-                component=component,
-                components=components,
-                attachment=attachment,
-                attachments=attachments,
-                embed=embed,
-                embeds=embeds,
-                replace_attachments=replace_attachments,
-                mentions_everyone=mentions_everyone,
-                user_mentions=user_mentions,
-                role_mentions=role_mentions,
-            )
-            return self._context._create_response(message)
-
-        message = await self._context.interaction.edit_initial_response(
-            content,
-            component=component,
-            components=components,
-            attachment=attachment,
-            attachments=attachments,
-            embed=embed,
-            embeds=embeds,
-            replace_attachments=replace_attachments,
-            mentions_everyone=mentions_everyone,
-            user_mentions=user_mentions,
-            role_mentions=role_mentions,
-        )
-        return self._context._create_response()
 
 
 # MIT License
